requirements.txt
huggingface
torch
torchvision
transformers
Pillow
ftfy
regex
tqdm